feat(rover): output service title for whoami (#299)
* feat(rover): output service title for whoami

In our discussions we found that Name was not a useful piece of
information to show to Rover's users. Instead, this commit introduces
a Service title which is aligned with the information that Studio
displays to a user. This commit ensures that Service title is only
printed for graph keys. For user keys we do not display this service
title.

Co-authored-by: Jake Dawkins <dawkinsjh@gmail.com>