infra: test, build and release on ubuntu-16.04
This pins us to Ubuntu 16.04 which ships with glib 2.19.  This should allow
us to work with a wider range of operating systems than the newer glib that
we get with Ubuntu 20.04, which is `ubuntu-latest` on GitHub Actions Virtual
Environments (which resulted in a Rover that wouldn't run on my Ubuntu 18.04).

Ubuntu 16.04 is LTS until April 2024, and is still receiving active updates
through the LTS program.