chore(deps): harmonizer@0.2.2
This brings in `harmonizer@0.2.2` which emits a core schema rather than
"CSDL".  The output of this is technically a breaking change for the `rover
supergraph compose` command because of the severity of CSDL -> Core but this
should brings stability to the format since the format itself is now
specified and governed by the core specification (whereas CSDL was just an
internal implementation that was unusable by consumers in its current form
within Rover anyway).