chore(rover_client): handle 400-599 HTTP responses (#395)
* chore(rover_client): handle 400-599 HTTP responses

When handling a response from an endpoint we are not currently handling
any unsuccessful requests.

For example, when a request returns a 405 response, we print out an
error saying that JSON could not be parsed. This happens because we
automatically try to parse the body of a response before checking if it
was successful or not. This commit adds that check.

This commit also does an extra check for 400 error, as the Apollo
Server sends additional information that we can display to the users.

* Apply suggestions from code review

Co-authored-by: Avery Harnish <avery@apollographql.com>

* rm DisabledIntrospection suggestion as it's no longer needed

Co-authored-by: Avery Harnish <avery@apollographql.com>