chore: begin adding shared types
I got a little carried away with this commit, but it's a good thing
I promise.

We've got lots of shared types now, I've got plans for a bunch more,
and things are looking cleaner than ever.

graph::check and subgraph::check were the primary drivers here,
but some other stuff was touched along the way.