sdl-encoder: encode string values with single quotes (")
We've been encoding all BlockStringValues and StringValues by surrounding them with `"""`. StringValue is meant to be encoded with just a single `"`. If a description does not include a line break or a `"`, we assume it's a StringValue.

This means that instead of printing this:
```graphql
"""Shopping list for cats at the pet store"""
type PetStoreTrip implements ShoppingTrip {
  toys: [DanglerPoleToys] @deprecated(reason: "Cats are too spoiled")
  """Dry or wet food?"""
  food: FoodType
  catGrass: Boolean
}
```

we encode this:
```graphql
"Shopping list for cats at the pet store"
type PetStoreTrip implements ShoppingTrip {
  toys: [DanglerPoleToys] @deprecated(reason: "Cats are too spoiled")
  "Dry or wet food?"
  food: FoodType
  catGrass: Boolean
}
```