Improve messaging for `rover subgraph check` (#980)
The current message outputted by rover when running a `subgraph check` only tells you if the API schema changed not the core schema.

This can sometimes lead to a confusing experience. For example, metadata directives such as `@key`, `@tag` etc. are removed during composition, so when they are added to the supergraph schema and there are no changes to the API schema, Rover tells the user that there were no changes, which is misleading.

This commit improves this behavior:

If the Core schema changed but the API schema did not change:

`There were no changes detected in the composed API schema, but the core schema was modified.`

If the Core and API schema did not change:

`There were no changes detected in the composed schema.`