fix(rover-client): introspection result should omit Float scalar
Since `Float` is a built-in GraphQL scalar type, it must not be part of the introspection results. This PR removes it from the SDL Rover produces. fixes #985

There is also a small tweak to 'interfaces' schema test to remove code examples within description, as it's not really necessary to test those specifics in an interface tests, and maintaining indentation is a bit of a pain!