Install rover before run if missing
Currently, rover installs its binary into `binary-install/node_modules/.bin`.
Changes to project dependencies followed by a `yarn` run wipe this file out,
supposedly with its caching mechanism. The `postinstall` script isn't rerun
when this happens, so the binary is left missing.

This updates the node bin to first check if the binary exists before running,
triggering an install if necessary. I'm not convinced this is a great
solution, but this is at least a significant improvement in UX for now.

Fixes #1178