feat: adds `--port` argument to `rover dev` (#1275)
fixes #1231 and #1266 by adding a `--port` to `rover dev` that configures the port that the router listens on.

each unique port creates its own `rover dev` session that will not interfere with routers running on other ports. additionally, if you try to run a subgraph on the same port as the router, you will get an error message letting you know, rather than the previous confusing message "there is a subgraph already running on that port".