chore: attempt multiple client writes at startup (#1336)
occasionally, the first `LocalSocketStream::connect` will incorrectly
succeed, indicating that the `rover dev` process should be an attached
process rather than a main process.

this PR changes the behavior to attempt 5 separate client connections
and writes/reads before determining the type of process this `rover dev`
should be. it may not eliminate 100% of these types of errors but should
cut down on them significantly.