chore: prebuilt ARM binaries for linux and macos (#1356)
This PR supercedes #1137 as the rebase was much too gnarly. This is
essentially that same PR. Copying the description w/updated status
below:

We should not merge this PR and release ARM support until we can run the
latest `rover supergraph compose` and `rover dev` commands on ARM.

The following must happen, in order, before landing this PR.

  - [x] Test ARM support with `federation-rs` release candidates
- [x] Release[`federation@2.0.6-rc.0`
_(javascript)_](https://github.com/apollographql/federation)
- [x] Release[`supergraph@2.0.6-rc.0`
_(rust)_](https://github.com/apollographql/federation-rs)
- [x] Test downloading and running `supergraph compose` with
`2.0.6-rc.0` in a `supergraph.yaml` with `rover` built from `main`
natively on an M1, and in Docker on an M1
- [x] Continue releasing release candidates until it all works.
~Hopefully it is just `2.0.6-rc.0`. If that works, it is safe to assume
0.x will work, so no release candidate will be needed there.~ _update:
it was not that easy, the next two steps are new_
- [x] Remove the CI builds for `aarch64-apple-darwin`
apollographql/federation-rs because [V8 snapshots cannot be
cross-compiled from x86_64
machines](https://github.com/denoland/deno/issues/4862#issuecomment-711110480)
and [CircleCI does not currently provide native ARM64 Mac
machines](https://circleci.com/docs/2.0/using-arm/#m1-mac-support)
meaning we _have_ to cross-compile if we want ARM binaries at all. If
either of those things change, we will revisit this decision.
- [x] Test downloading and running `supergraph compose` with
`2.0.6-rc.1` in a `supergraph.yaml` with `rover` built from `main`
natively on an arm ubuntu device, and in Docker on an M1
- [x] Release first stable `federation-rs` releases with native ARM
support (only Linux binaries, no ARM64 MacOS binaries)
- [x] Release [`federation@2.1.0`
_(javascript)_](https://github.com/apollographql/federation)
- [x] Release [`supergraph@2.1.0`
_(rust)_](https://github.com/apollographql/federation-rs)
- [x] Release [`federation@0.37.0`
_(javascript)_](https://github.com/apollographql/federation)
- [x] Release [`supergraph@0.37.0`
_(rust)_](https://github.com/apollographql/federation-rs)
- [x] Test downloading and running `supergraph compose` with the
packages we just released
- [x] Release [`router@1.1.0`](https://github.com/apollographql/router)
  - [x] Prepare ARM support for Rover
    - [x] Cut `rover@v0.9.2-rc.0` 
- [x] Test downloading and running `supergraph compose` natively on an
M1, and in Docker on an M1
    - [x] Release more rcs until we are confident everything works
    - [x] Wait for the cloud router to default to version 2.1.x