chore: dont use registry url in build-rs (#1407)
sometimes you may want to run a command against staging from a fresh
build of rover, in this case you could run

`APOLLO_REGISTRY_URL="https://staging.url" cargo rover config whoami`.
Unfortunately this would cause the schema to fail to fetch because
`build.rs` was reading `APOLLO_REGISTRY_URL` to fetch the build schema,
and that schema doesn't exist in staging. we want the schema to come
from prod, and we can send the request wherever we want.