feat: Add `rover contract describe` and `rover contract publish` (#1475)
This PR implements the first phase of
https://github.com/apollographql/rover/issues/1421

That is, this PR:
- Implements `rover contract describe`, which prints out a
human-readable description of the configuration of a contract variant.

```
Describe the configuration of a contract variant from the Apollo graph registry

Usage: rover contract describe [OPTIONS] <GRAPH_REF>

Arguments:
  <GRAPH_REF>
          <NAME>@<VARIANT> of graph in Apollo Studio. @<VARIANT> may be left off, defaulting to @current
```

- Implements `rover contract publish`, which allows creating a contract
variant (or updating its configuration) and optionally triggering a
launch.

```
Publish an updated contract configuration to the Apollo graph registry and trigger launch in the graph router

Usage: rover contract publish [OPTIONS] <--include-tag <INCLUDE_TAG>|--no-include-tags> <--exclude-tag <EXCLUDE_TAG>|--no-exclude-tags> <--hide-unreachable-types|--no-hide-unreachable-types> <GRAPH_REF>

Arguments:
  <GRAPH_REF>
          <NAME>@<VARIANT> of graph in Apollo Studio. @<VARIANT> may be left off, defaulting to @current

Options:
      --source-variant <SOURCE_VARIANT>
          The source variant name for this contract variant. Once set, this cannot be changed

      --include-tag <INCLUDE_TAG>
          List of tag names to include in the contract schema (e.g. '--include-tag foo --include-tag bar'). To specify an empty list, use --no-include-tags instead

      --no-include-tags
          Use an empty include list of tag names for the contract schema. To specify a non-empty list, use --include-tag instead

      --exclude-tag <EXCLUDE_TAG>
          List of tag names to exclude from the contract schema (e.g. '--exclude-tag foo --exclude-tag bar'). To specify an empty list, use --no-exclude-tags instead

      --no-exclude-tags
          Use an empty exclude list of tag names for the contract schema. To specify a non-empty list, use --exclude-tag instead

      --hide-unreachable-types
          Automatically hide types that can never be reached in operations on the contract schema

      --no-hide-unreachable-types
          Do not automatically hide types that can never be reached in operations on the contract schema

      --no-launch
          Do not trigger a launch in Studio after updating the contract configuration
```

Some notes:
- The API for upserting a contract's configuration does not allow
specifying a particular key as `null` to indicate "don't change this
key's value". This means that most configuration is required to be
specified each time the command is run.
- The exception is `--source-variant`, which the API allows to be
optional.
- Follow-up work will take on making these configuration keys nullable
in the API. I've set up the Rover code such that making the switch
shouldn't be too much work.
- Should mainly be changing the `ArgGroup`s to use `required(false)`,
updating the `runner.rs`/`types.rs`, and removing the `None` check in
`run()`.
- Note that instead of saying "not specifying `--include` or
`--no-include` means use the pre-existing value", we could also
introduce a new argument e.g. `--keep-include` that explicitly keeps the
pre-existing value, and add `--keep-include` to the `ArgGroup` instead
of making the `ArgGroup` not required. (Similar situation for the other
configuration options.) This is something to consider in the follow-up
work though/doesn't need to be considered here.

Co-authored-by: Caydie Tran <caydie@apollographql.com>
Co-authored-by: Avery Harnish <avery@apollographql.com>