release: v0.11.0-rc.2 (#1482)
This release includes some changes to `rover dev`:

1) There is now a `--router-config <ROUTER_YAML_PATH>` argument that you
can pass to specify a yaml configuration file to start the router with.
This file is reloaded when it is changed.
2) `rover dev` will now intelligently select a version of the router and
composition to use based on the versions in
[`./latest_plugin_versions.json`](https://github.com/apollographql/rover/blob/main/latest_plugin_versions.json)
3) `-n` is now an alias for `--name`

There are two new commands: `rover contract publish` and `rover contract
describe`. Docs on these commands can be found
[here](https://www.apollographql.com/docs/rover/commands/contracts).

The `--output` argument has been renamed to `--format` and should be
used for specifying JSON output, and `--output` should be used for
writing Rover's command output to a file.