release: v0.13.0 (#1539)Co-authored-by: Dylan Anthony <dbanty@users.noreply.github.com>
# [0.13.0] - 2023-03-10

## 🚀 Features

- **Dynamic templates - @dbanty, #1388**

Rover now fetches templates from a GraphQL API. New templates can be
contributed to [the templates
repo](https://github.com/apollographql/templates) and will not require a
new release of Rover to be used. To see the latest templates, run `rover
template list`.

## 🐛 Fixes

- **Fixes behavior of `--router-config` in `rover dev` -
@EverlastingBugstopper, #1536 fixes #1535**

Older versions of Rover did not respect the `supergraph` section of the
`--router-config` file. This has been fixed in v0.13.0.

---------

Co-authored-by: Dylan Anthony <dbanty@users.noreply.github.com>