feat: migrate `rover config whoami` output to stdout and add json format (#1560)
## Description

BREAKING CHANGE: This pr fixes #1380 by using a RoverOutput `whoami`
variant to print the output of the `rover config whoami` command to
stdout instead of stderr. It also adds the capability to print the
output as json with the `--json` flag.
This PR migrates the output of rover config whoami command from stderr
to stdout and fixes #1380. Additionally, this PR adds the --json flag to
allow users to retrieve the output in JSON format. This change improves
the usability of the rover CLI and enhances the user experience.

## Related Issues & discussions
- #1380 Rover config whoami should print to stdout instead of stderr and
return JSON data
- #60 Design consistent output printing

### Related PRs
- #68 standardize command return type