fix: removes duplicated `IsFederatedQuery` and allows `--convert` (#1575)
This PR removes the `IsFederatedQuery` check from the routing URL
fetcher. This enables the flow where you are trying to publish a
subgraph for the first time to a variant that used to be a monograph.
Additionally, it removes the need for the `IsFederatedQuery` to be sent
twice on each `rover subgraph publish`.