feat: warn when `--routing-url` is unroutable (#1565)
Supersedes #1543 (took a different approach, didn't care to destroy /
overwrite the history there)

Fixes https://github.com/apollographql/rover/issues/1477

This feature addition adds URL validation to the rover subgraph publish
`--routing-url` flag. It also introduces a dependent flag to bypass
warnings/prompts `--allow-invalid-routing-url`.

Taken from @EverlastingBugstopper 's
https://github.com/apollographql/rover/issues/1477#issuecomment-1397506739:
When a `--routing-url` is passed to rover subgraph publish:

In the case where a human is manually doing a publish (this can be
determined by checking for a TTY) then Rover prints a warning that the
URL is invalid with a prompt (`[y/N]`) to continue with the publish.

In the case where CI or a script is doing the publish (check for a TTY)
then Rover prints a warning saying the routing url is invalid, but still
allow the publish to go through. We might also want to make the warning
include a notice that this behavior may break in future versions of
Rover if the `--allow-invalid-routing-url` flag is not passed.

In both scenarios (and in future versions of Rover), passing the
`--allow-invalid-routing-url` flag should disable the warnings entirely
and allow the invalid routing url to be used.

When a `--routing-url` is not passed to rover subgraph publish:

Perform a GraphQL query against the platform API to get the routing url
for that subgraph
If that URL is invalid, follow the same error/warn procedure as
described above

<!--
First, 🌠 thank you 🌠 for taking the time to consider a contribution to
Apollo!

Here are some important details to follow:

* ⏰ Your time is important
To save your precious time, if the contribution you are making will
take more than an hour, please make sure it has been discussed in an
        issue first. This is especially true for feature requests!

* 💡 Features
Feature requests can be created and discussed within a GitHub Issue.
Be sure to search for existing feature requests (and related issues!)
prior to opening a new request. If an existing issue covers the need,
please upvote that issue by using the 👍 emote, rather than opening a
        new issue.

* 🕷 Bug fixes
These can be created and discussed in this repository. When fixing a
bug,
please _try_ to add a test which verifies the fix. If you cannot, you
should
still submit the PR but we may still ask you (and help you!) to create a
test.

* 📖 Contribution guidelines
Follow https://github.com/apollographql/rover/blob/HEAD/CONTRIBUTING.md
when submitting a pull request. Make sure existing tests still pass, and
add
        tests for all new behavior.

* ✏️ Explain your pull request
Describe the big picture of your changes here to communicate to what
        your pull request is meant to accomplish. Provide 🔗 links 🔗 to
        associated issues!

We hope you will find this to be a positive experience! Open source
contribution can be intimidating and we hope to alleviate that pain as
much
as possible. Without following these guidelines, you may be missing
context
that can help you succeed with your contribution, which is why we
encourage
discussion first. Ultimately, there is no guarantee that we will be able
to
merge your pull-request, but by following these guidelines we can try to
avoid disappointment.

-->

---------

Co-authored-by: Avery Harnish <avery@apollographql.com>