pq publish: remove preview phase message (#1686)
The feature will be moving from private preview to public preview soon,
and you will no longer need the feature to be enabled on your GraphOS
account. And until that happens, an attempt to use it on a graph without
the feature activated will still get the following error from GraphOS
servers:

error: Exception while fetching data (/_entities[0]/persistedQueryList) :
Persisted Query Lists are a Preview-stage feature which is not enabled
for graph GRAPH. If you are interested in this upcoming functionality,
    please reach out to support@apollographql.com.

So we don't need this additional unconditional message.