feat(install): support configs for binary mirrors (#1675)
This adds the ability to install the rover binary utilizing npm in
situations where a custom registry is defined and you sit behind a
outgoing firewall/proxy. In these situations, organizations typically
have some sort of artifact repository acting as a proxy/mirror for npm
and internet endpoints, and users need to be able to specify what proxy
to pull the binary from to react `rover.apollo.dev`.

(This is the same way popular packages that have binaries work, such as
Playwright, Cypress, Canvas, Electron, etc...)

This PR adds the ability to specify via npm configs, the apollo rover
download host. (see updates to getting-started.md)

---------

Co-authored-by: Maria Elisabeth Schreiber <maria.schreiber@apollographql.com>