fix: point lint diagnostics to the right place when using unicode
Ariadne expects to get character offsets (as in the `str::chars()`
iterator), while the lint service returns byte offsets into the UTF-8
input string. This patch maps byte offsets to character offsets before
passing them to ariadne.

The mapping is as simple as possible, so it does take a lot of memory
(one `usize` per character, so 8x the input file size). It could be
reduced to 4x by using `u32` if large files are a concern?

This should not be necessary in the future after we patch ariadne, but
that is a big refactor.