deps: inline binary-install, with updated axios version (#1819)
There is currently an open security warning on axios versions < 1.0,
used by binary-install. Users were never affected by this issue (it's
only relevant for cross-site requests, which installation packages never
do), but they do get scary warnings when they install rover.

There are some open PRs on binary-install that address this but it may
be easier to fix it by copying the code into the rover installer. I
updated axios to 1.0. We can also remove the `rimraf` dependency as its
functionality is built-in to all our target versions of Node.js
(14.14+).

It seems like axios 1.0 has some proxy environment related code built in
now, but I opted to keep the `axios-proxy-builder` package so we don't
need to spend time comparing the two.