Use task specific `rayon` threadpools and not the global threadpool
The global `rayon` threadpool has a variable number of threads:
 - An environment variable can be used to specify the number: RAYON_NUM_THREADS
 - If not set, The number of CPUs on the executing system is used

Either of these may lead to a situation where the global threadpool does
not have enough threads for `rover dev` to execute correctly. It's not
realistic to expect customers to know how to set this environment
variable and a better option is available.

At each point in `rover` where we use `rayon`, create a thread pool which
is task specific and explicitly sets the number of threads available to
the pool.

For example:

```
    let tp = rayon::ThreadPoolBuilder::new()
        .num_threads(1)
        .thread_name(|idx| format!("router-runner-{idx}"))
        .build()
        .map_err(|err| {
            RoverError::new(anyhow!("could not create router runner thread pool: {err}",))
        })?;
```
This has the added advantage of allowing us to provide a name for all
the threads created by the pool which is helpful debugging information.

Note: It might be that a more comprehensive answer to this problem is to
not use `rayon` at all, but that's a bigger change than I want to
consider for fixing this specific issue.

fixes: #1871