fix(rover-std): avoid panics on file watcher errors (#1935)
Remove panics on file watcher errors, and display better error messages
when rover fails to watch a file.

The `panic!` macro was used in the `watch_file` method, because it
spawns a rayon ThreadPool to monitor the file out of band.

This PR changes the logic in three way:

1. adds extra `eprintln!` with useful information when facing errors;
2. breaks from loops or early returns when facing an error that invoked
a `panic!` before; and
3. change the signature of the crossbeam channel to send a `Result<(),
RoverStdError>` instead of `()`, so we can send the error back to the
initial caller.

There's one error pattern that's unhandled, but safe. When failing to
send a message to the crossbeam channel, it fails silently. Since this
only occurs when the channel is closed, this likely means that the
calling thread stopped working.