chore(ci+deps): Fix up Release CI and Explicitly add `tokio` feature flag (#1972)
Now that we've added in new and extra tests the `release` job in CI needed updating to account for these. In addition while preparing an abandoned release it was realised we needed to explicitly tell `tokio` to use the `rt-multi-thread` feature flag, instead of inheriting it from other packages we also use.