Ensure we handle, branches being empty for CircleCI (#1981)
CircleCI's variable for GitHub branch is left empty when pushing a tag
so our GHA tests fail when we try and run them for a release. To this
end, make xtask a little more resilient to this, preferring `main` if we
don't push a specific branch with the idea that our tagging 99.9% of the
time will always be to the head commit of `main`.