Don't run IsFederatedGraph before running SubgraphFetchQuery (#2004)
All the checks done by IsFederatedGraph are also done by
SubgraphFetchQuery (checking to see if graph/variant exists and to see
if the graph is federated). This is just an additional round-trip adding
latency and potential errors. Additionally, with `rover supergraph
compose` which is commonly run against many subgraphs of a single graph,
these will run fully redundantly on every subgraph that is fetched.