release: v0.26.0-rc.0 (#2009)
This release candidate contains changes to how `rover` queries the
Apollo Platform API to reduce its strain on the system. This should lead
to shorter runtimes for composition and other similar operations that
fetch subgraphs. Testing should therefore focus on these areas.

Additionally adds the "BSD-2-Clause" licence to the allow list, as we
presently allow the "BSD-3-Clause" licence and the former is a subset of
the latter, we already support this licence type we are simply making it
explicit.