Add `--graph-ref` to `supergraph compose` (#2001)
**NOTE: This should not be merged until #1874 has been merged, and then
this rebased on top**

In #1984 we added support to pass a `--graph-ref` flag to `rover dev`.
Since `supergraph compose` is performing the same function as `rover
dev` it makes sense to expose the same option there too. However, this
requires some refactoring in order to make work.

In principle the idea is fairly simple, we should centralise the
construction of the SupergraphConfig object so that both `dev` and
`supergraph compose` can hook in to it from a central location. However
to do this I've also had to add a fair amount of testing because there
was very little coverage in this area, however, due to @dotdat's work on
#1977 we have a lot more test coverage than before. That work has been
integrated into this PR.

Further more @dotdat was working on another PR #1985 , which optimises
the calls made by the `--graph-ref` option to fetch subgraphs in one
request rather than multiple. Due to being in this area and working on
the `--graph-ref` option it also made sense to integrate that PR here as
well, to prevent a complicated rebase later on.

This PR has been tested by the addition of unit tests and running
`supergraph compose` locally.

_There is also a small change to `deny.toml` in this PR to add support
for BSD-2-Clause licences. We already support BSD-3-Clause, and 2-Clause
is strictly a subset so this should not be a problem_