ROVER-83 Add tests for `whoami` (#2022)
In this PR I've added a test for `whoami` and also consolidated some of the other 'integration-like' tests into the End-To-End test framework. This is pointing towards an eventual demarkation where we have end-to-end tests and more traditional integration tests sitting side by side but this is gradual process that we'll move towards over time.