feat(tests): Let E2Es message if scheduled run contains failures (#2033)
Now we have E2E tests that run nightly, it would be much better to let
them message us in Slack if there's a failure in the run. This PR
implements that, by computing an aggregated status for the matrix run,
passing it out of the re-usable workflow and then acting in the calling
workflow.