Tightening up how we do Slack messaging
Turns out we can use GitHub Actions features to do this better. If we
remove `continue-on-error`, and replace it by turning off 'fail-fast',
we get all the smoke tests to run which gives us all the errors, and
the re-usable workflow gets a nice status.

We can then track the status in the calling workflow, and execute
the Slack message on the back of that. Which is much nicer.