Add step to fail workflow if matrix branch fails (#2044)
As comment indicates, GitHub Actions handling of matrices and the
overall workflow result appears to be quite janky and not at all as
you'd expect. So this adds a final job that will only run if we get
failures, this will then cause the workflow to fail which will bubble up
into the calling workflow.